chore: extract base skeleton
